split the UT code into a ut file.
Because UT is compiled and linked as a separate step.
