+++ btree/Makefile.sub
@@ -1,7 +1,5 @@
-nobase_motr_include_HEADERS += \
-                            btree/btree.h \
-                            btree/internal.h
+nobase_motr_include_HEADERS += btree/btree.h \
+                               btree/internal.h
 
-motr_libmotr_la_SOURCES  += \
-                            btree/btree.c
+motr_libmotr_la_SOURCES  += btree/btree.c
