+++ btree/btree.c
@@ -517,7 +517,6 @@
 #include "module/instance.h"
 #include "lib/memory.h"
 #include "lib/assert.h"
-#include "ut/ut.h"          /** struct m0_ut_suite */
 //#include "lib/user_space/misc.h" /** ARRAY_SIZE_FOR_BITS */
 
 #ifndef __KERNEL__
@@ -1024,92 +1023,92 @@ static uint64_t         trees_in_use[ARRAY_SIZE_FOR_BITS(M0_TREE_COUNT,
 static uint32_t         trees_loaded = 0;
 static struct m0_rwlock trees_lock;
 
-static bool node_invariant(const struct nd *node)
+static bool node__invariant(const struct nd *node)
 {
 	return node->n_type->nt_invariant(node);
 }
 
 static int node_count(const struct nd *node)
 {
-	M0_PRE(node_invariant(node));
+	M0_PRE(node__invariant(node));
 	return node->n_type->nt_count(node);
 }
 
 static int node_space(const struct nd *node)
 {
-	M0_PRE(node_invariant(node));
+	M0_PRE(node__invariant(node));
 	return node->n_type->nt_space(node);
 }
 
 #if 0
 static int node_level(const struct nd *node)
 {
-	M0_PRE(node_invariant(node));
+	M0_PRE(node__invariant(node));
 	return (node->n_type->nt_level(node));
 }
 
 static int node_shift(const struct nd *node)
 {
-	M0_PRE(node_invariant(node));
+	M0_PRE(node__invariant(node));
 	return (node->n_type->nt_shift(node));
 }
 
 static void node_fid(const struct nd *node, struct m0_fid *fid)
 {
-	M0_PRE(node_invariant(node));
+	M0_PRE(node__invariant(node));
 	node->n_type->nt_fid(node, fid);
 }
 #endif
 
 static void node_rec(struct slot *slot)
 {
-	M0_PRE(node_invariant(slot->s_node));
+	M0_PRE(node__invariant(slot->s_node));
 	slot->s_node->n_type->nt_rec(slot);
 }
 
 #if 0
 static void node_key(struct slot *slot)
 {
-	M0_PRE(node_invariant(slot->node));
+	M0_PRE(node__invariant(slot->node));
 	slot->s_node->n_type->nt_key(slot);
 }
 
 static void node_child(struct slot *slot, struct segaddr *addr)
 {
-	M0_PRE(node_invariant(slot->node));
+	M0_PRE(node__invariant(slot->node));
 	slot->s_node->n_type->nt_child(slot, addr);
 }
 #endif
 
 static bool node_isfit(struct slot *slot)
 {
-	M0_PRE(node_invariant(slot->s_node));
+	M0_PRE(node__invariant(slot->s_node));
 	return slot->s_node->n_type->nt_isfit(slot);
 }
 
 static void node_done(struct slot *slot, struct m0_be_tx *tx, bool modified)
 {
-	M0_PRE(node_invariant(slot->s_node));
+	M0_PRE(node__invariant(slot->s_node));
 	slot->s_node->n_type->nt_done(slot, tx, modified);
 }
 
 static void node_make(struct slot *slot, struct m0_be_tx *tx)
 {
-	M0_PRE(node_invariant(slot->s_node));
+	M0_PRE(node__invariant(slot->s_node));
 	slot->s_node->n_type->nt_make(slot, tx);
 }
 
 #ifndef __KERNEL__
-static void node_find(struct slot *slot, const struct m0_btree_key *key)
+M0_UNUSED static void node_find(struct slot *slot, const struct m0_btree_key *key)
 {
-	M0_PRE(node_invariant(slot->s_node));
+	M0_PRE(node__invariant(slot->s_node));
 	slot->s_node->n_type->nt_find(slot, key);
 }
 #endif
 
 static void node_fix(const struct nd *node, struct m0_be_tx *tx)
 {
-	M0_PRE(node_invariant(node));
+	M0_PRE(node__invariant(node));
 	node->n_type->nt_fix(node, tx);
 }
 
@@ -1117,14 +1116,14 @@ static void node_fix(const struct nd *node, struct m0_be_tx *tx)
 static void node_cut(const struct nd *node, int idx, int size,
 		     struct m0_be_tx *tx)
 {
-	M0_PRE(node_invariant(node));
+	M0_PRE(node__invariant(node));
 	node->n_type->nt_cut(node, idx, size, tx);
 }
 #endif
 
 static void node_del(const struct nd *node, int idx, struct m0_be_tx *tx)
 {
-	M0_PRE(node_invariant(node));
+	M0_PRE(node__invariant(node));
 	node->n_type->nt_del(node, idx, tx);
 }
 
@@ -1132,8 +1131,8 @@ static void node_del(const struct nd *node, int idx, struct m0_be_tx *tx)
 static void node_move(struct nd *src, struct nd *tgt,
 		      enum dir dir, int nr, struct m0_be_tx *tx)
 {
-	M0_PRE(node_invariant(src));
-	M0_PRE(node_invariant(dst));
+	M0_PRE(node__invariant(src));
+	M0_PRE(node__invariant(dst));
 	M0_IN(dir,(D_LEFT, D_RIGHT));
 	tgt->n_type->nt_move(src, tgt, dir, nr, tx);
 }
@@ -1154,7 +1153,7 @@ struct mod {
 	const struct m0_btree_type *m_ttype[TTYPE_NR];
 };
 
-int m0_btree_mod_init(void)
+M0_INTERNAL int m0_btree_mod_init(void)
 {
 	struct mod *m;
 
@@ -1171,7 +1170,7 @@ int m0_btree_mod_init(void)
 		return M0_ERR(-ENOMEM);
 }
 
-void m0_btree_mod_fini(void)
+M0_INTERNAL void m0_btree_mod_fini(void)
 {
 	m0_rwlock_fini(&trees_lock);
 	m0_free(mod_get());
@@ -1355,7 +1354,7 @@ static int64_t tree_get(struct node_op *op, struct segaddr *addr, int nxt)
  *
  * @return Next state to proceed in.
  */
-static int64_t tree_create(struct node_op *op, struct m0_btree_type *tt,
+M0_UNUSED static int64_t tree_create(struct node_op *op, struct m0_btree_type *tt,
 			   int rootshift, struct m0_be_tx *tx, int nxt)
 {
 	return segops->so_tree_create(op, tt, rootshift, tx, nxt);
@@ -1371,7 +1370,7 @@ static int64_t tree_create(struct node_op *op, struct m0_btree_type *tt,
  *
  * @return Next state to proceed in.
  */
-static int64_t tree_delete(struct node_op *op, struct td *tree,
+M0_UNUSED static int64_t tree_delete(struct node_op *op, struct td *tree,
 			   struct m0_be_tx *tx, int nxt)
 {
 	M0_PRE(tree != NULL);
@@ -1468,7 +1467,7 @@ static int64_t node_alloc(struct node_op *op, struct td *tree, int size,
 }
 
 #ifndef __KERNEL__
-static int64_t node_free(struct node_op *op, struct nd *node,
+M0_UNUSED static int64_t node_free(struct node_op *op, struct nd *node,
 			 struct m0_be_tx *tx, int nxt)
 {
 	node->n_type->nt_fini(node);
@@ -2040,344 +2039,6 @@ static void generic_move(struct nd *src, struct nd *tgt,
 }
 
 
-#ifndef __KERNEL__
-/**
- * The code contained below is 'ut'. This is a little experiment to contain the
- * ut code in the same file containing the functionality code. We are open to
- * changes iff enough reasons are found that this model either does not work or
- * is not intuitive or maintainable.
- */
-static void m0_btree_ut_node_create_delete(void);
-static void m0_btree_ut_node_add_del_rec(void);
-
-
-/**
- * btree_ut test suite.
- */
-struct m0_ut_suite btree_ut = {
-	.ts_name = "btree-ut",
-	.ts_yaml_config_string = "{ valgrind: { timeout: 3600 },"
-	"  helgrind: { timeout: 3600 },"
-	"  exclude:  ["
-	"   "
-	"  ] }",
-	.ts_init = NULL,
-	.ts_fini = NULL,
-	.ts_tests = {
-		{"node_create_delete",          m0_btree_ut_node_create_delete},
-		{"node_add_del_rec",            m0_btree_ut_node_add_del_rec},
-		{NULL, NULL}
-	}
-};
-
-static bool btree_ut_initialised = false;
-static void btree_ut_init(void)
-{
-
-	if (!btree_ut_initialised) {
-		segops = (struct seg_ops *)&mem_seg_ops;
-		m0_rwlock_init(&trees_lock);
-		btree_ut_initialised = true;
-	}
-}
-
-static void btree_ut_fini(void)
-{
-	segops = NULL;
-	m0_rwlock_fini(&trees_lock);
-	btree_ut_initialised = false;
-}
-
-/**
- * This test will create a few nodes and then delete them before exiting. The
- * main intent of this test is to debug the create and delete nodes functions.
- */
-static void m0_btree_ut_node_create_delete(void)
-{
-	struct node_op          op;
-	struct m0_btree_type    tt;
-	struct td              *tree;
-	struct td              *tree_clone;
-	struct nd              *node1;
-	struct nd              *node2;
-	const struct node_type *nt    = &fixed_format;
-
-	M0_ENTRY();
-
-	btree_ut_init();
-
-	M0_SET0(&op);
-
-	M0_ASSERT(trees_loaded == 0);
-
-	// Create a Fixed-Format tree.
-	op.no_opc = NOP_ALLOC;
-	tree_create(&op, &tt, 10, NULL, 0);
-
-	tree = op.no_tree;
-
-	M0_ASSERT(tree->r_ref == 1);
-	M0_ASSERT(tree->t_root != NULL);
-	M0_ASSERT(trees_loaded == 1);
-
-	// Add a few nodes to the created tree.
-	op.no_opc = NOP_ALLOC;
-	node_alloc(&op, tree, 10, nt, 8, 8, NULL, 0);
-	node1 = op.no_node;
-
-	op.no_opc = NOP_ALLOC;
-	node_alloc(&op,  tree, 10, nt, 8, 8, NULL, 0);
-	node2 = op.no_node;
-
-	op.no_opc = NOP_FREE;
-	node_free(&op, node1, NULL, 0);
-
-	op.no_opc = NOP_FREE;
-	node_free(&op, node2, NULL, 0);
-
-	/* Get another reference to the same tree. */
-	tree_get(&op, &tree->t_root->n_addr, 0);
-	tree_clone = op.no_tree;
-	M0_ASSERT(tree_clone->r_ref == 2);
-	M0_ASSERT(tree->t_root == tree_clone->t_root);
-	M0_ASSERT(trees_loaded == 1);
-
-
-	tree_put(tree_clone);
-	M0_ASSERT(trees_loaded == 1);
-
-	// Done playing with the tree - delete it.
-	op.no_opc = NOP_FREE;
-	tree_delete(&op, tree, NULL, 0);
-	M0_ASSERT(trees_loaded == 0);
-
-	btree_ut_fini();
-	M0_LEAVE();
-}
-
-
-static bool add_rec(struct nd *node,
-		    uint64_t   key,
-		    uint64_t   val)
-{
-	struct ff_head      *h = ff_data(node);
-	struct slot          slot;
-	struct m0_btree_key  find_key;
-	m0_bcount_t          ksize;
-	void                *p_key;
-	m0_bcount_t          vsize;
-	void                *p_val;
-
-	/**
-	 * To add a record if space is available in the node to hold a new
-	 * record:
-	 * 1) Search index in the node where the new record is to be inserted.
-	 * 2) Get the location in the node where the key & value should be
-	 *    inserted.
-	 * 3) Insert the new record at the determined location.
-	 */
-
-	ksize = h->ff_ksize;
-	p_key = &key;
-	vsize = h->ff_vsize;
-	p_val = &val;
-
-	M0_SET0(&slot);
-	slot.s_node                            = node;
-	slot.s_rec.r_key.k_data.ov_vec.v_nr    = 1;
-	slot.s_rec.r_key.k_data.ov_vec.v_count = &ksize;
-	slot.s_rec.r_val.ov_vec.v_nr           = 1;
-	slot.s_rec.r_val.ov_vec.v_count        = &vsize;
-
-	if (node_count(node) != 0) {
-		if (!node_isfit(&slot))
-			return false;
-		find_key.k_data.ov_vec.v_nr = 1;
-		find_key.k_data.ov_vec.v_count = &ksize;
-		find_key.k_data.ov_buf = &p_key;
-		node_find(&slot, &find_key);
-	}
-
-	node_make(&slot, NULL);
-
-	slot.s_rec.r_key.k_data.ov_buf = &p_key;
-	slot.s_rec.r_val.ov_buf = &p_val;
-
-	node_rec(&slot);
-
-	*((uint64_t *)p_key) = key;
-	*((uint64_t *)p_val) = val;
-
-	return true;
-}
-
-static void get_next_rec_to_add(struct nd *node, uint64_t *key,  uint64_t *val)
-{
-	struct slot          slot;
-	uint64_t             proposed_key;
-	struct m0_btree_key  find_key;
-	m0_bcount_t          ksize;
-	void                *p_key;
-	m0_bcount_t          vsize;
-	void                *p_val;
-	struct ff_head      *h = ff_data(node);
-
-	M0_SET0(&slot);
-	slot.s_node = node;
-
-	ksize = h->ff_ksize;
-	proposed_key = rand();
-
-	find_key.k_data.ov_vec.v_nr    = 1;
-	find_key.k_data.ov_vec.v_count = &ksize;
-	find_key.k_data.ov_buf         = &p_key;
-
-	slot.s_rec.r_key.k_data.ov_vec.v_nr = 1;
-	slot.s_rec.r_key.k_data.ov_vec.v_count = &ksize;
-	slot.s_rec.r_key.k_data.ov_buf = &p_key;
-
-	slot.s_rec.r_val.ov_vec.v_nr = 1;
-	slot.s_rec.r_val.ov_vec.v_count = &vsize;
-	slot.s_rec.r_val.ov_buf = &p_val;
-	while (true) {
-		uint64_t found_key;
-
-		proposed_key %= 256;
-		p_key = &proposed_key;
-
-		if (node_count(node) == 0)
-			break;
-		node_find(&slot, &find_key);
-		node_rec(&slot);
-
-		if (slot.s_idx >= node_count(node))
-			break;
-
-		found_key = *(uint64_t *)p_key;
-
-		if (found_key == proposed_key)
-			proposed_key++;
-		else
-			break;
-	}
-
-	*key = proposed_key;
-	memset(val, *key, sizeof(*val));
-}
-
-void get_rec_at_index(struct nd *node, int idx, uint64_t *key,  uint64_t *val)
-{
-	struct slot          slot;
-	m0_bcount_t          ksize;
-	void                *p_key;
-	m0_bcount_t          vsize;
-	void                *p_val;
-
-	M0_SET0(&slot);
-	slot.s_node = node;
-	slot.s_idx  = idx;
-
-	M0_ASSERT(idx<node_count(node));
-
-	slot.s_rec.r_key.k_data.ov_vec.v_nr = 1;
-	slot.s_rec.r_key.k_data.ov_vec.v_count = &ksize;
-	slot.s_rec.r_key.k_data.ov_buf = &p_key;
-
-	slot.s_rec.r_val.ov_vec.v_nr = 1;
-	slot.s_rec.r_val.ov_vec.v_count = &vsize;
-	slot.s_rec.r_val.ov_buf = &p_val;
-
-	node_rec(&slot);
-
-	if (key != NULL)
-		*key = *(uint64_t *)p_key;
-
-	if (val != NULL)
-		*val = *(uint64_t *)p_val;
-}
-
-/**
- * This test will create a tree, add a node and then populate the node with
- * some records. It will also confirm the records are in ascending order of Key.
- */
-void m0_btree_ut_node_add_del_rec(void)
-{
-	struct node_op          op;
-	struct m0_btree_type    tt;
-	struct td              *tree;
-	struct nd              *node1;
-	const struct node_type *nt      = &fixed_format;
-	uint64_t                key;
-	uint64_t                val;
-	uint64_t                prev_key;
-	uint64_t                curr_key;
-	time_t                  curr_time;
-	int                     run_loop;
-
-	M0_ENTRY();
-
-	time(&curr_time);
-	printf("\nUsing seed %lu", curr_time);
-	srand(curr_time);
-
-	run_loop = 50000;
-
-	btree_ut_init();
-
-	M0_SET0(&op);
-
-	op.no_opc = NOP_ALLOC;
-	tree_create(&op, &tt, 10, NULL, 0);
-
-	tree = op.no_tree;
-
-	op.no_opc = NOP_ALLOC;
-	node_alloc(&op, tree, 10, nt, 8, 8, NULL, 0);
-	node1 = op.no_node;
-
-	while (run_loop--) {
-		int i;
-
-		/** Add records */
-		i = 0;
-		while (true) {
-			get_next_rec_to_add(node1, &key, &val);
-			if (!add_rec(node1, key, val))
-				break;
-			M0_ASSERT(++i == node_count(node1));
-		}
-
-		/** Confirm all the records are in ascending value of key. */
-		get_rec_at_index(node1, 0, &prev_key, NULL);
-		for (i = 1; i < node_count(node1); i++) {
-			get_rec_at_index(node1, i, &curr_key, NULL);
-			M0_ASSERT(prev_key < curr_key);
-			prev_key = curr_key;
-		}
-
-		/** Delete all the records from the node. */
-		i = node_count(node1) - 1;
-		while (node_count(node1) != 0) {
-			int j = rand() % node_count(node1);
-			node_del(node1, j, NULL);
-			M0_ASSERT(i-- == node_count(node1));
-		}
-	}
-
-	printf("\n");
-	op.no_opc = NOP_FREE;
-	node_free(&op, node1, NULL, 0);
-
-	// Done playing with the tree - delete it.
-	op.no_opc = NOP_FREE;
-	tree_delete(&op, tree, NULL, 0);
-
-	btree_ut_fini();
-
-	M0_LEAVE();
-}
-
-#endif  /** KERNEL */
 #undef M0_TRACE_SUBSYSTEM
 
 
@@ -2387,6 +2048,10 @@ void m0_btree_ut_node_add_del_rec(void)
  * - test how cookies affect performance (for large trees and small trees);
  */
 /** @} end of btree group */
+
+/*
+ * LocalWords:  btree allocator smop smops
+ */
 
 /*
  *  Local variables:
@@ -2399,7 +2064,4 @@ void m0_btree_ut_node_add_del_rec(void)
  */
 /*
  * vim: tabstop=8 shiftwidth=8 noexpandtab textwidth=80 nowrap
- */
-
-/*  LocalWords:  btree allocator smop smops
  */
