+++ btree/ut/Makefile.sub
@@ -1,0 +1,1 @@
+ut_libmotr_ut_la_SOURCES += btree/ut/btree_ut.c
